refactor(gridexplore-app): Add a container component which manage notification websockets to centralize treeview data managment.
